backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.11",
    "django-cors-headers==4.3.1",
    "django-extensions==3.2.3",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "pillow==10.2.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.1",
]
